+++ .storybook/preview.tsx
@@ -10,7 +10,10 @@ import { SBProvider } from "./components/SBProvider";
 import { StoryDetails } from "./components/StoryDetails";
 
 import "@jobber/design/foundation.css";
+import "@jobber/design/dark.mode.css";
 import "./assets/css/preview.css";
+import { Unstyled } from "@storybook/addon-docs";
+import { Table } from "@storybook/components";
 
 export const parameters = {
   viewMode: "docs",
@@ -44,7 +47,7 @@ export const parameters = {
   },
   docs: {
     components: {
-      wrapper: props => <Content>{props.children}</Content>,
+      wrapper: props => <Unstyled><Content>{props.children}</Content></Unstyled>,
       h1: props => <Header {...props} level={1} />,
       h2: props => <Header {...props} level={2} isTOC />,
       h3: props => <Header {...props} level={3} />,
@@ -52,6 +55,7 @@ export const parameters = {
       h5: props => <Header {...props} level={5} />,
       p: props => <Text {...props} />,
       inlineCode: props => <InlineCode {...props} />,
+      table: props => <Table {...props} />,
       Canvas: CustomCanvas,
     },
     container: DocsWithSidebar,
@@ -60,9 +64,15 @@ export const parameters = {
 
 export const decorators = [
   Story => (
-    <SBProvider>
-      <Story />
-    </SBProvider>
+    <div style={{ backgroundColor: 'var(--color-surface)' }}>
+      <SBProvider>
+        <Story />
+      </SBProvider>
+    </div>
   ),
-  (Story, context) => <StoryDetails Story={Story} context={context} />,
+  (Story, context) =>
+    <div style={{ backgroundColor: 'var(--color-surface)' }}>
+      <StoryDetails Story={Story} context={context} />
+    </div>
+  ,
 ];