+++ packages/components/src/InputAvatar/InputAvatar.tsx
@@ -11,7 +11,7 @@ interface InputAvatarProps extends Omit<AvatarProps, "size"> {
    * to upload the file.
    *
    * More info is available at:
-   * https://atlantis.getjobber.com/?path=/docs/components-forms-and-inputs-inputfile-docs--page#getuploadparams
+   * https://atlantis.getjobber.com/?path=/docs/components-forms-and-inputs-inputfile--docs#getuploadparams
    */
   getUploadParams(file: File): UploadParams | Promise<UploadParams>;